refactor: NlpToken 클래스들 token 패키지로 이동
