+++ src/main/java/com/almondia/meca/cardhistory/infra/morpheme/token/NlpToken.java
@@ -1,4 +1,4 @@
-package com.almondia.meca.cardhistory.infra.morpheme;
+package com.almondia.meca.cardhistory.infra.morpheme.token;
 
 public interface NlpToken {
